Unnecessary cast from char to int
git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/lang/trunk@430168 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/org/apache/commons/lang/StringEscapeUtilsTest.java
@@ -256,8 +256,8 @@ public class StringEscapeUtilsTest extends TestCase {
             Character c1 = new Character(i);
             Character c2 = new Character((char)(i+1));
             String expected = c1.toString() + c2.toString();
-            String escapedC1 = "&#x" + Integer.toHexString((int)(c1.charValue())) + ";";
-            String escapedC2 = "&#x" + Integer.toHexString((int)(c2.charValue())) + ";";
+            String escapedC1 = "&#x" + Integer.toHexString((c1.charValue())) + ";";
+            String escapedC2 = "&#x" + Integer.toHexString((c2.charValue())) + ";";
             assertEquals("hex number unescape index " + (int)i, expected, StringEscapeUtils.unescapeHtml(escapedC1 + escapedC2));
         }
     }
